More advanced cache clearing
The previous variant of cache clearing required the cache to pass over
each response.  A structure has been put in place to minimize the search
through this list.  This approach will be slower in simple cases, but
should be benificial in the situation where the cache has a lot of
items.