Add support for clear keys, format code and optional debug logging
An endpoint was added to directly send clear
keys.  (/.mu/clear-keys).  This endpoint reads the clear-keys header
and removes keys from the endpoint.

Logging has been added by virtue of the `LOG_CACHE_KEYS` and
`LOG_CLEAR_KEYS` environment variables.  Setting these to "on" or
"true" emits the optional logging.  This is especially handy when
debugging the cache clearing.