More robust handling of clear_keys useage with fix and cleanup
When the `/.mu/clear-keys` request was sent and the CLEAR_KEYS were
not being logged, the clear keys would erroneously not be passed
through.  This has been solved.

In the same go, we have abstracted the logging of CACHE_KEYS and
CLEAR_KEYS making the methods easier to understand.

We have also abstracted the code so it would cope with multiple
CLEAR_KEYS headers being set in the same request to
`/.mu/clear-keys`.  Albeit not necessary the changes make the code
easier to reason about.