A rough PoC of request folding
This PoC shows requests could be folded.  Connections aren't all stopped
correctly, the status code isn't taken into account, interrupting
clear_keys aren't taken into a account, and a broken main receiver will
likely break all the consumers.

A full fledged implementation might be better served without the use of
plug_mint_proxy, building directly on top of mint.  It may also be that
the plug_mint_proxy could be extended so we'd only use the forwarding
and interfacing parts of it.  Either case, this PoC is a good showcase
of what can be.